refactor: JanggiGame의 클래스명을 Board로 변경
